Move ServiceInfo to tracer-api within custom tracer submodule.
