+++ apm-agent-plugins/apm-servlet-plugin/src/main/java/co/elastic/apm/agent/servlet/ServletServiceNameHelper.java
@@ -18,8 +18,8 @@
  */
 package co.elastic.apm.agent.servlet;
 
-import co.elastic.apm.agent.configuration.ServiceInfo;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
+import co.elastic.apm.agent.tracer.service.ServiceAwareTracer;
+import co.elastic.apm.agent.tracer.service.ServiceInfo;
 import co.elastic.apm.agent.sdk.logging.Logger;
 import co.elastic.apm.agent.sdk.logging.LoggerFactory;
 import co.elastic.apm.agent.sdk.state.GlobalState;
@@ -44,8 +44,8 @@ public class ServletServiceNameHelper {
                                                              @Nullable ServletContext servletContext,
                                                              Tracer tracer) {
 
-        ElasticApmTracer elasticApmTracer = tracer.probe(ElasticApmTracer.class);
-        if (elasticApmTracer == null) {
+        ServiceAwareTracer serviceAwareTracer = tracer.probe(ServiceAwareTracer.class);
+        if (serviceAwareTracer == null) {
             return;
         }
 
@@ -63,7 +63,7 @@ public class ServletServiceNameHelper {
             return;
         }
         ServiceInfo serviceInfo = detectServiceInfo(adapter, servletContext, servletContextClassLoader);
-        elasticApmTracer.setServiceInfoForClassLoader(servletContextClassLoader, serviceInfo);
+        serviceAwareTracer.setServiceInfoForClassLoader(servletContextClassLoader, serviceInfo);
     }
 
     public static <ServletContext> ServiceInfo detectServiceInfo(ServletContextAdapter<ServletContext> adapter, ServletContext servletContext, ClassLoader servletContextClassLoader) {
