+++ apm-agent-tracer/src/main/java/co/elastic/apm/agent/tracer/service/ServiceInfo.java
@@ -16,33 +16,21 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.configuration;
-
-import co.elastic.apm.agent.sdk.internal.util.PrivilegedActionUtils;
+package co.elastic.apm.agent.tracer.service;
 
 import javax.annotation.Nullable;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Properties;
 import java.util.jar.Attributes;
-import java.util.jar.JarFile;
 import java.util.jar.Manifest;
 
 public class ServiceInfo {
-
-    private static final String JAR_VERSION_SUFFIX = "-(\\d+\\.)+(\\d+)(.*)?$";
     private static final String DEFAULT_SERVICE_NAME = "unknown-java-service";
     private static final ServiceInfo EMPTY = new ServiceInfo(null, null);
-    private static final ServiceInfo AUTO_DETECTED = autoDetect(System.getProperties(), PrivilegedActionUtils.getEnv());
 
     private final String serviceName;
     @Nullable
     private final String serviceVersion;
     private final boolean multiServiceContainer;
-
-    public ServiceInfo(@Nullable String serviceName) {
-        this(serviceName, null);
-    }
 
     private ServiceInfo(@Nullable String serviceName, @Nullable String serviceVersion) {
         this(serviceName, serviceVersion, false);
@@ -82,76 +70,6 @@ public class ServiceInfo {
         return serviceName.replaceAll("[^a-zA-Z0-9 _-]", "-");
     }
 
-    public static ServiceInfo autoDetected() {
-        return AUTO_DETECTED;
-    }
-
-    public static ServiceInfo autoDetect(Properties sysProperties, Map<String,String> sysEnv) {
-        String lambdaFunctionName = sysEnv.get("AWS_LAMBDA_FUNCTION_NAME");
-        if (lambdaFunctionName != null) {
-            return new ServiceInfo(lambdaFunctionName, sysEnv.get("AWS_LAMBDA_FUNCTION_VERSION"));
-        } else {
-            ServiceInfo serviceInfo = createFromSunJavaCommand(sysProperties.getProperty("sun.java.command"));
-            if (serviceInfo != null) {
-                return serviceInfo;
-            }
-            return ServiceInfo.empty();
-        }
-    }
-
-    @Nullable
-    private static ServiceInfo createFromSunJavaCommand(@Nullable String command) {
-        if (command == null) {
-            return null;
-        }
-        command = command.trim();
-        String serviceName = getContainerServiceName(command);
-        if (serviceName != null) {
-            return ServiceInfo.ofMultiServiceContainer(serviceName);
-        }
-        if (command.contains(".jar")) {
-            return fromJarCommand(command);
-        } else {
-            return fromMainClassCommand(command);
-        }
-    }
-
-    @Nullable
-    private static String getContainerServiceName(String command) {
-        if (command.startsWith("org.apache.catalina.startup.Bootstrap")) {
-            return "tomcat-application";
-        } else if (command.startsWith("org.eclipse.jetty")) {
-            return "jetty-application";
-        } else if (command.startsWith("com.sun.enterprise.glassfish")) {
-            return "glassfish-application";
-        } else if (command.contains("ws-server.jar")) {
-            return "websphere-application";
-        } else if (command.contains("jboss-modules.jar")) {
-            return "jboss-application";
-        } else if (command.contains("weblogic")) {
-            return "weblogic-application";
-        }
-        return null;
-    }
-
-    private static ServiceInfo fromJarCommand(String command) {
-        final String[] commandParts = command.split(" ");
-        ServiceInfo serviceInfoFromManifest = ServiceInfo.empty();
-        ServiceInfo serviceInfoFromJarName = ServiceInfo.empty();
-        for (String commandPart : commandParts) {
-            if (commandPart.endsWith(".jar")) {
-                try (JarFile jarFile = new JarFile(commandPart)) {
-                    serviceInfoFromManifest = fromManifest(jarFile.getManifest());
-                } catch (Exception ignored) {
-                }
-
-                serviceInfoFromJarName = ServiceInfo.of(removeVersionFromJar(removePath(removeJarExtension(commandPart))));
-                break;
-            }
-        }
-        return serviceInfoFromManifest.withFallback(serviceInfoFromJarName);
-    }
-
     public static ServiceInfo fromManifest(@Nullable Manifest manifest) {
         if (manifest == null) {
             return ServiceInfo.empty();
@@ -160,29 +78,6 @@ public class ServiceInfo {
         return ServiceInfo.of(
             mainAttributes.getValue(Attributes.Name.IMPLEMENTATION_TITLE),
             mainAttributes.getValue(Attributes.Name.IMPLEMENTATION_VERSION));
-    }
-
-    private static String removeJarExtension(String commandPart) {
-        return commandPart.substring(0, commandPart.indexOf(".jar"));
-    }
-
-    private static String removePath(String path) {
-        return path.substring(path.lastIndexOf("/") + 1).substring(path.lastIndexOf("\\") + 1);
-    }
-
-    private static String removeVersionFromJar(String jarFileName) {
-        return jarFileName.replaceFirst(JAR_VERSION_SUFFIX, "");
-    }
-
-    private static ServiceInfo fromMainClassCommand(String command) {
-        final String mainClassName;
-        int indexOfSpace = command.indexOf(' ');
-        if (indexOfSpace != -1) {
-            mainClassName = command.substring(0, indexOfSpace);
-        } else {
-            mainClassName = command;
-        }
-        return new ServiceInfo(mainClassName.substring(mainClassName.lastIndexOf('.') + 1));
     }
 
     public String getServiceName() {
