+++ integration-tests/application-server-integration-tests/src/test/java/co/elastic/apm/servlet/tests/ExternalPluginTestApp.java
@@ -18,7 +18,7 @@
  */
 package co.elastic.apm.servlet.tests;
 
-import co.elastic.apm.agent.configuration.ServiceInfo;
+import co.elastic.apm.agent.tracer.service.ServiceInfo;
 import co.elastic.apm.agent.impl.Tracer;
 import co.elastic.apm.agent.tracer.Outcome;
 import co.elastic.apm.servlet.AbstractServletContainerIntegrationTest;
